+++ init.rb
@@ -1,1 +1,0 @@
-require 'tabs_on_rails'