+++ lib/tabs_on_rails/action_controller.rb
@@ -3,7 +3,7 @@
 #
 # A simple Ruby on Rails plugin for creating and managing Tabs.
 #
-# Copyright (c) 2009-2013 Simone Carletti <weppos@weppos.net>
+# Copyright (c) 2009-2017 Simone Carletti <weppos@weppos.net>
 #++
 
 
@@ -103,7 +103,7 @@ module TabsOnRails
         options = args.extract_options!
         name, namespace = args
 
-        before_filter(options) do |controller|
+        before_action(options) do |controller|
           controller.send(:set_tab, name, namespace)
         end
       end