+++ lib/tabs_on_rails/railtie.rb
@@ -3,19 +3,16 @@
 #
 # A simple Ruby on Rails plugin for creating and managing Tabs.
 #
-# Copyright (c) 2009-2013 Simone Carletti <weppos@weppos.net>
+# Copyright (c) 2009-2017 Simone Carletti <weppos@weppos.net>
 #++
 
 
 module TabsOnRails
 
   class Railtie < Rails::Railtie
-    initializer "tabs_on_rails.initialize" do
+    ActiveSupport.on_load(:action_controller) do
+      ::ActionController::Base.send(:include, TabsOnRails::ActionController)
     end
   end
 
-end
-
-ActiveSupport.on_load(:action_controller) do
-  include TabsOnRails::ActionController
 end