+++ test/dummy.rb
@@ -8,7 +8,7 @@ require "rails/railtie"
 class Dummy
   Routes = ActionDispatch::Routing::RouteSet.new
   Routes.draw do
-    match ':controller(/:action(/:id))'
+    match ':controller(/:action(/:id))', via: [:get]
   end
 end
 
@@ -39,3 +39,8 @@ class ActiveSupport::TestCase
   end
 
 end
+
+# Trigger lazy loading and causes the load_hooks to be executed on
+# ActionController::API. This is important because breacrumbs_on_rails includes
+# BreadcrumbsOnRails::ActionController on any module that executes these hooks
+ActionController::API rescue NameError